+++ src/hardis-status-provider.ts
@@ -37,6 +37,7 @@ export class HardisStatusProvider
    */
   private async getTopicElements(topic: any): Promise<StatusTreeItem[]> {
     const items: StatusTreeItem[] = [];
+    console.debug("Starting TreeViewItem_init_" + topic.id + " ...");
     console.time("TreeViewItem_init_" + topic.id);
     const topicItems: any[] =
       topic.id === "status-org"
@@ -149,6 +150,27 @@ export class HardisStatusProvider
           vscode.window.showWarningMessage(item.tooltip);
         }
         items.push(item);
+      }
+      // Scratch org pool info
+      if (options.devHub) {
+        const poolViewRes = await execSfdxJson(
+          "sfdx hardis:scratch:pool:view",
+          this,
+          { output: false, fail: false }
+        );
+        if (
+          poolViewRes?.status === 0 &&
+          (poolViewRes?.result?.availableScratchOrgs ||
+            poolViewRes?.result?.availableScratchOrgs === 0)
+        ) {
+          items.push({
+            id: "scratch-org-pool-view",
+            label: `Pool: ${poolViewRes.result.availableScratchOrgs} available (max ${poolViewRes.result.maxScratchOrgs})`,
+            tooltip: "View content of scratch org pool",
+            command: "sfdx hardis:scratch:pool:view",
+            icon: "pool.svg",
+          });
+        }
       }
       items.push({
         id: "select-another-org" + (options.devHub ? "-devhub" : ""),